Install ALE for Vim
ALE is an Asynchronous Linting Engine which works with Vim 8 and NeoVim
to lint code "as you type". It does so without blocking the main thread
by using the asynchronous features of both of those Vim versions.

The asynchronous nature means linting does not block other interactions
with Vim and it's been nice to get linting feedback more often than
buffer write.

The "as-you-type" linting is performant enough to actually be a
distraction, so this tones that down. With this configuration, ALE will
lint on cursor hold durations of 1 second and when leaving insert mode
in addition to buffer write, buffer read, and file type change.