dotfiles,laptop.sh: don't use ASDF for Go
Not using it is simpler, one less thing to debug.

ASDF is particularly useful when there are packages that break when a
language is upgraded. This happens sometimes in languages like Ruby and
JavaScript, where it is more useful to have multiple versions of the
language installed, but very rare in Go.

In large part, that is because of Go's compatibility promise.
https://golang.org/doc/go1compat

If there were a package that works on one version of Go but breaks in
the next, that may be a reason to stop using that package.

When trying out a beta release, Go has a convenient way to do that:

```
$ go get https://golang.org/dl/go1.11beta2
$ go1.11beta2 download
...
$ go1.11beta2 version
go version go1.11beta2 linux/amd64
```

That has the benefit that the resulting command is called
`go1.11beta2` not just `go`,
so it's explicit that you're using a beta,
and there's no need to "switch" to it or back again.