dotfiles/vim: set encoding to UTF-8
```
:h encoding

For GTK+ 2 or later, it is highly recommended to set 'encoding'
to "utf-8".  Although care has been taken to allow different values of
'encoding', "utf-8" is the natural choice for the environment and
avoids unnecessary conversion overhead.  "utf-8" has not been made
the default to prevent different behavior of the GUI and terminal
versions, and to avoid changing the encoding of newly created files
without your knowledge (in case 'fileencodings' is empty).
```