dotfiles: improve "find references for word under cursor"
When editing a file type that supports Language Server Protocol
(such as TypeScript),
map `fr` to `:ALEFindReference` to improve accuracy.

Otherwise, use `:grep`, which is mapped to `ag` (The Silver Searcher).

`ag` picks up on `gitignore` to ignore certain file patterns.
Add `dist` to the list of ignored patterns to improve search quality
in projects that use the Parcel bundler.

https://parceljs.org/