dotfiles/git: ignore .ignore files in Git
There are some directories in my current primary monorepo
that I don't want to search with `ag`.

I want to ignore those files for searches,
but keep them in version control
and not leave an unnecessary artifact in the repo for teammates
since this is a local development environment concern for me.