bin/kill-pid-running: add
Similar to the other script `kill-pid-on-port`,
this kills a running process with a given name.

I've needed this recently as I've been trying
`sqls`, a language server for SQL that connects
to my app's database in order to provide cool
in-editor stuff. But then, when I want to do a
`pg_restore`, there's a connection open,
which I needed to find and kill.
This does it in one step.