vim: run_file with split_cmd in a TTY
Programs like `mods` often check if their standard output (stdout) is
connected to a terminal (TTY). When I run a command using `:!` in
Neovim, the stdout is not a TTY, so `mods` disables color output. This
is why I see color when running the command directly in my terminal but
not when running it from within Neovim using `:!`.

To fix this, I modified my `run_file` function to run the command inside
a Neovim terminal buffer instead of using `:!`. When I use a terminal
buffer, stdout is connected to a TTY, and `mods` will output colorized
text as expected.