vim/ai: use goose instead of mods
Goose has the ability to edit files, run tests, etc.
The term of art in the industry is to "use tools".

https://block.github.io/goose/

I like that Goose is open source, written in Rust, maintained by the
Block team, and built upon the Model Context Protocol. These feel like
solid foundations upon which to build our tools.

This change installs and configures Goose:

- Update laptop.sh to install the Goose CLI tool via curl.
- Append Goose installation directory to PATH in shell/zshrc.
- Extend Vim configuration with:
  - A new function (`run_goose`) to run Goose commands in a split terminal.
  - Key mappings (`<Leader>r` for starting a new session and `<Leader>c`
    for resuming) that leverage the Goose CLI.
  - An `Esc` key mapping easily exit terminal insert mode.
  - An `autocmd` to set the `filetype` for `.goosehints` files to Markdown.